CBO.ZERO can also be used on uncacheable or not-cached memory locations
Some CMO operations, such as CBO.ZERO, can be used on memory addresses
that are not cached or which are marked as uncacheable, even if this
is not their primary use case.  Expressly permit CBO.ZERO to be used
in these situations, and adjust the language elsewhere in the
specification to permit this.